chore(migrations) cassandra to always connect migrations and use the returned coordinator
### Summary

In current migrations with Cassandra there are `connect:query` and `coordinator:execute`
calls, with no obvious reason to what use and when. This just causes a little bit of
confusion so I decided to always `connect_migrations` and use the returned coordinator
to execute actual migrations rather than using `connector:query` in some cases.

All in all, `connect_migrations` stores the coordinator, so end result is almost 1:1,
except with `connector:query` there is a wrapper function for `execute`, which I didn't
find we need for migrations.