fix(db) cassandra connector wait_for_schema_consensus to wait full timeout even on errors
### Summary

It was reported by @jeremyjpj0916 on #6403 that Cassandra migrations might fail with:

```
[Cassandra error] cluster_mutex callback threw an error: migrations.lua:142:
[Cassandra error] failed to wait for schema consensus: [Read failure] Operation failed -
                  received 0 responses and 1 failures
```

The timestamps that Jeremy posted showed us that before failure this got logged:

```
2020/09/28 17:20:14 [verbose] waiting for Cassandra schema consensus (120000ms timeout)...
2020/09/28 17:20:14 [verbose] Cassandra schema consensus: not reached
```

See, the code was set to wait for schema consensus for 120 seconds, yet the code seemed
to fail right away. Thus it means the code in this case didn't try to reach schema
consensus for 120 secs as the log line might indicate. This PR changes the wait so
that we try it for full 120 secs and only fail if it didn't reach the consensus
even after that. We try to reach consensus every 100 ms until the timeout occurs
with the `max_schema_consensus_wait` setting that is configurable with:
`cassandra_schema_consensus_timeout` configuration value.

### Issues Resolved

Should fix the remaining of #6403